Hey Jerrryyy
I think this is everything from a code point of view that needed to change in the code but can you check just to make sure. the "fourbar" is now being called the "lift" becuase its not only a 4 bar now and we also have a fourbar. the "chainbar" is now called "fourbar" bacuase thats what it is now. it has two potentiometers not just one. also the wheel base is on the power expander not the lift anymore so that changed some of the functions. we still need to tune all the pid and everything but i think this is all the structure changes needed